+++ .github/scripts/affected-modules.js
@@ -1,0 +1,97 @@
+function globToRegex(glob) {
+  // Split on ** so single-segment expansion of * doesn't see them.
+  const escapeSegment = (seg) =>
+    seg.replace(/[.+?^${}()|[\]\\]/g, "\\$&").replace(/\*/g, "[^/]*");
+  const body = glob.split("**").map(escapeSegment).join(".*");
+  return new RegExp(`^${body}$`);
+}
+
+/**
+ * Returns module-graph helpers for a given (elements, rules) config.
+ */
+function createAffectedModules(elements, rules) {
+  const compiledElements = elements.map((el) => ({
+    type: el.type,
+    regex: globToRegex(el.pattern),
+  }));
+
+  const allTypes = elements.map((e) => e.type);
+
+  function expandPattern(pattern) {
+    if (pattern === "*") {
+      return [...allTypes];
+    }
+    if (pattern.endsWith("/*")) {
+      const prefix = pattern.slice(0, -1);
+      return allTypes.filter((t) => t.startsWith(prefix));
+    }
+    return allTypes.includes(pattern) ? [pattern] : [];
+  }
+
+  // Inverse adjacency: dependentsOf.get(M) is the set of modules that may
+  // import M (i.e. modules that need to be invalidated when M changes).
+  const dependentsOf = new Map(allTypes.map((t) => [t, new Set()]));
+  for (const rule of rules) {
+    const fromTypes = rule.from.flatMap(expandPattern);
+    const allowTypes = rule.allow.flatMap(expandPattern);
+    for (const target of allowTypes) {
+      for (const importer of fromTypes) {
+        if (importer !== target) {
+          dependentsOf.get(target).add(importer);
+        }
+      }
+    }
+  }
+
+  function fileToModule(path) {
+    for (const el of compiledElements) {
+      if (el.regex.test(path)) {
+        return el.type;
+      }
+    }
+    return null;
+  }
+
+  function directlyTouchedModules(changedFiles) {
+    const direct = new Set();
+    for (const file of changedFiles) {
+      const m = fileToModule(file);
+      if (m) {
+        direct.add(m);
+      }
+    }
+    return direct;
+  }
+
+  function affectedModules(changedFiles) {
+    const direct = directlyTouchedModules(changedFiles);
+    const affected = new Set(direct);
+    const queue = [...direct];
+    while (queue.length > 0) {
+      const m = queue.shift();
+      for (const dep of dependentsOf.get(m) ?? []) {
+        if (!affected.has(dep)) {
+          affected.add(dep);
+          queue.push(dep);
+        }
+      }
+    }
+    return affected;
+  }
+
+  function selectTests(affected, testFiles) {
+    return testFiles.filter((f) => {
+      const m = fileToModule(f);
+      return m !== null && affected.has(m);
+    });
+  }
+
+  return {
+    fileToModule,
+    directlyTouchedModules,
+    affectedModules,
+    selectTests,
+  };
+}
+
+module.exports = { createAffectedModules, globToRegex };
