+++ .github/workflows/frontend.yml
@@ -12,7 +12,7 @@ on:
       skip-custom-viz:
         type: boolean
         default: false
-      # JSON array of unit-test paths produced by compute-test-selection in
+      # JSON array of unit-test paths produced by compute-test-plan in
       # run-tests.yml. Currently unused — will drive Jest --testPathPatterns
       # once module-boundary violations reach zero.
       unit-tests-to-run:
